requirements.txt
# to read environments if set up
python-dotenv

# Pick one of the libraries. Ldap3 seems to have less dependencies
ldap3

#connection libraries. Only cx-oracle is version bound
cx-oracle==8.3.0